refactor: Merge TypeScript type files
