+++ packages/react-native-reanimated/src/animationBuilder.tsx
@@ -3,8 +3,8 @@ import type {
   ILayoutAnimationBuilder,
   LayoutAnimationFunction,
   LayoutAnimationsValues,
-} from './layoutReanimation';
-import type { StyleProps } from './commonTypes';
+  StyleProps,
+} from './commonTypes';
 import type { NestedArray } from './createAnimatedComponent/commonTypes';
 import { logger } from './logger';
