+++ packages/react-native-reanimated/src/component/FlatList.tsx
@@ -9,9 +9,8 @@ import type {
 import { FlatList } from 'react-native';
 import { AnimatedView } from './View';
 import { createAnimatedComponent } from '../createAnimatedComponent';
-import type { ILayoutAnimationBuilder } from '../layoutReanimation/animationBuilder/commonTypes';
+import type { AnimatedStyle, ILayoutAnimationBuilder } from '../commonTypes';
 import { LayoutAnimationConfig } from './LayoutAnimationConfig';
-import type { AnimatedStyle } from '../commonTypes';
 import type { AnimatedProps } from '../helperTypes';
 
 const AnimatedFlatList = createAnimatedComponent(FlatList);
