+++ packages/react-native-reanimated/src/createAnimatedComponent/createAnimatedComponent.tsx
@@ -14,7 +14,8 @@ import invariant from 'invariant';
 import { adaptViewConfig } from '../ConfigHelper';
 import { RNRenderer } from '../platform-specific/RNRenderer';
 import { enableLayoutAnimations } from '../core';
-import { SharedTransition, LayoutAnimationType } from '../layoutReanimation';
+import { SharedTransition } from '../layoutReanimation';
+import { LayoutAnimationType } from '../commonTypes';
 import type { StyleProps, ShadowNodeWrapper } from '../commonTypes';
 import { getShadowNodeWrapperFromRef } from '../fabricUtils';
 import { removeFromPropsRegistry } from '../PropsRegistry';
