+++ packages/react-native-reanimated/src/layoutReanimation/animationBuilder/BaseAnimationBuilder.ts
@@ -1,12 +1,12 @@
 'use strict';
 import { withDelay } from '../../animation';
-import type {
-  EntryExitAnimationFunction,
-  AnimationFunction,
-  LayoutAnimationFunction,
-} from './commonTypes';
 
 import { ReduceMotion } from '../../commonTypes';
+import type {
+  AnimationFunction,
+  EntryExitAnimationFunction,
+  LayoutAnimationFunction,
+} from '../../commonTypes';
 import { getReduceMotionFromConfig } from '../../animation/util';
 import { ReanimatedError } from '../../errors';
