+++ packages/react-native-reanimated/src/layoutReanimation/animationBuilder/index.ts
@@ -2,23 +2,3 @@
 export { BaseAnimationBuilder } from './BaseAnimationBuilder';
 export { ComplexAnimationBuilder } from './ComplexAnimationBuilder';
 export { Keyframe } from './Keyframe';
-export { LayoutAnimationType, SharedTransitionType } from './commonTypes';
-export type {
-  LayoutAnimation,
-  AnimationFunction,
-  EntryAnimationsValues,
-  ExitAnimationsValues,
-  EntryExitAnimationFunction,
-  AnimationConfigFunction,
-  IEntryAnimationBuilder,
-  IExitAnimationBuilder,
-  LayoutAnimationsValues,
-  LayoutAnimationFunction,
-  LayoutAnimationStartFunction,
-  ILayoutAnimationBuilder,
-  BaseLayoutAnimationConfig,
-  BaseBuilderAnimationConfig,
-  LayoutAnimationAndConfig,
-  IEntryExitAnimationBuilder,
-  SharedTransitionAnimationsValues,
-} from './commonTypes';
