+++ packages/react-native-reanimated/src/layoutReanimation/defaultAnimations/Bounce.ts
@@ -1,10 +1,10 @@
 'use strict';
+import { withSequence, withTiming } from '../../animation';
 import type {
-  EntryExitAnimationFunction,
   EntryExitAnimationsValues,
   IEntryExitAnimationBuilder,
-} from '../animationBuilder/commonTypes';
-import { withSequence, withTiming } from '../../animation';
+  EntryExitAnimationFunction,
+} from '../../commonTypes';
 import type { BaseAnimationBuilder } from '../animationBuilder';
 import { ComplexAnimationBuilder } from '../animationBuilder';
