+++ packages/react-native-reanimated/src/layoutReanimation/defaultAnimations/Fade.ts
@@ -2,7 +2,7 @@
 import type {
   IEntryExitAnimationBuilder,
   EntryExitAnimationFunction,
-} from '../animationBuilder/commonTypes';
+} from '../../commonTypes';
 import type { BaseAnimationBuilder } from '../animationBuilder';
 import { ComplexAnimationBuilder } from '../animationBuilder';
