+++ packages/react-native-reanimated/src/layoutReanimation/defaultAnimations/Lightspeed.ts
@@ -1,13 +1,12 @@
 'use strict';
 import { withSequence, withTiming } from '../../animation';
+import type {
+  IEntryExitAnimationBuilder,
+  EntryExitAnimationFunction,
+  EntryExitAnimationsValues,
+} from '../../commonTypes';
 import type { BaseAnimationBuilder } from '../animationBuilder';
 import { ComplexAnimationBuilder } from '../animationBuilder';
-import type {
-  EntryExitAnimationsValues,
-  EntryExitAnimationFunction,
-  IEntryExitAnimationBuilder,
-} from '../animationBuilder/commonTypes';
-
 /**
  * Entry from right animation with change in skew and opacity. You can modify
  * the behavior by chaining methods like `.springify()` or `.duration(500)`.
