+++ packages/react-native-reanimated/src/layoutReanimation/defaultAnimations/Roll.ts
@@ -1,11 +1,11 @@
 'use strict';
+import type {
+  IEntryExitAnimationBuilder,
+  EntryExitAnimationFunction,
+  EntryExitAnimationsValues,
+} from '../../commonTypes';
 import type { BaseAnimationBuilder } from '../animationBuilder';
 import { ComplexAnimationBuilder } from '../animationBuilder';
-import type {
-  EntryExitAnimationsValues,
-  EntryExitAnimationFunction,
-  IEntryExitAnimationBuilder,
-} from '../animationBuilder/commonTypes';
 
 /**
  * Roll from left animation. You can modify the behavior by chaining methods
