+++ packages/react-native-reanimated/src/layoutReanimation/defaultAnimations/Zoom.ts
@@ -2,13 +2,13 @@
 import type {
   IEntryExitAnimationBuilder,
   EntryExitAnimationFunction,
-  EntryAnimationsValues,
-  ExitAnimationsValues,
   EntryExitAnimationsValues,
-  AnimationConfigFunction,
   IEntryAnimationBuilder,
+  AnimationConfigFunction,
+  EntryAnimationsValues,
   IExitAnimationBuilder,
-} from '../animationBuilder/commonTypes';
+  ExitAnimationsValues,
+} from '../../commonTypes';
 import type { BaseAnimationBuilder } from '../animationBuilder';
 import { ComplexAnimationBuilder } from '../animationBuilder';
