+++ packages/react-native-reanimated/src/layoutReanimation/defaultTransitions/SequencedTransition.ts
@@ -3,7 +3,7 @@ import { withSequence, withTiming } from '../../animation';
 import type {
   ILayoutAnimationBuilder,
   LayoutAnimationFunction,
-} from '../animationBuilder/commonTypes';
+} from '../../commonTypes';
 import { BaseAnimationBuilder } from '../animationBuilder';
 
 /**
