+++ packages/react-native-reanimated/src/layoutReanimation/index.ts
@@ -4,4 +4,3 @@ export * from './animationBuilder';
 export * from './defaultAnimations';
 export * from './defaultTransitions';
 export * from './sharedTransitions';
-export type { KeyframeProps } from './animationBuilder/commonTypes';
