+++ packages/react-native-reanimated/src/layoutReanimation/sharedTransitions/SharedTransition.ts
@@ -1,18 +1,18 @@
 'use strict';
 import { withTiming } from '../../animation';
 import type {
+  CustomProgressAnimation,
+  LayoutAnimationsOptions,
+  ProgressAnimation,
   SharedTransitionAnimationsFunction,
   SharedTransitionAnimationsValues,
-  CustomProgressAnimation,
-  ProgressAnimation,
-  LayoutAnimationsOptions,
-} from '../animationBuilder/commonTypes';
+  StyleProps,
+} from '../../commonTypes';
 import {
   LayoutAnimationType,
+  ReduceMotion,
   SharedTransitionType,
-} from '../animationBuilder/commonTypes';
-import type { StyleProps } from '../../commonTypes';
-import { ReduceMotion } from '../../commonTypes';
+} from '../../commonTypes';
 import { ProgressTransitionManager } from './ProgressTransitionManager';
 import { updateLayoutAnimations } from '../../UpdateLayoutAnimations';
 import { getReduceMotionFromConfig } from '../../animation/util';
