+++ packages/react-native-reanimated/src/layoutReanimation/web/componentUtils.ts
@@ -13,10 +13,9 @@ import type { WebEasingsNames } from './Easing.web';
 import type { TransitionData } from './animationParser';
 import { TransitionGenerator } from './createAnimation';
 import { scheduleAnimationCleanup } from './domUtils';
+import { LayoutAnimationType, ReduceMotion } from '../../commonTypes';
 import { _updatePropsJS } from '../../js-reanimated';
 import type { ReanimatedHTMLElement } from '../../js-reanimated';
-import { ReduceMotion } from '../../commonTypes';
-import { LayoutAnimationType } from '../animationBuilder/commonTypes';
 import type { ReanimatedSnapshot, ScrollOffsets } from './componentStyle';
 import { setElementPosition, snapshots } from './componentStyle';
 import { Keyframe } from '../animationBuilder';
